7 Holmefield Avenue
"It is, I think, much more likely that Henry Baker is an absolutely innocent man, who had no idea
that the bird which he was carrying was of considerably more value than if it were made of solid
gold. That, however, I shall determine by a very simple test if we have an answer to our
advertisement."

"And you can do nothing until then?"

"Nothing."

"In that case I shall continue my professional round. But I shall come back in the evening at the
hour you have mentioned, for I should like to see the solution of so tangled a business."

"Very glad to see you. I dine at seven. There is a woodcock, I believe. By the way, in view of
recent occurrences, perhaps I ought to ask Mrs. Hudson to examine its crop."

I had been delayed at a case, and it was a little after half-past six when I found myself in Baker
Street once more. As I approached the house I saw a tall man in a Scotch bonnet with a coat which
was buttoned up to his chin waiting outside in the bright semicircle which was thrown from the
fanlight. Just as I arrived the door was opened, and we were shown up together to Holmes' room.