125 Harvey Street
"Yes, it was only put there a couple of years ago."

"Your sister asked for it, I suppose?"

"No, I never heard of her using it. We used always to get what we wanted for ourselves."

"Indeed, it seemed unnecessary to put so nice a bell-pull there. You will excuse me for a few
minutes while I satisfy myself as to this floor." He threw himself down upon his face with his lens
in his hand and crawled swiftly backward and forward, examining minutely the cracks between the
boards. Then he did the same with the wood-work with which the chamber was panelled. Finally he
walked over to the bed and spent some time in staring at it and in running his eye up and down the
wall. Finally he took the bell-rope in his hand and gave it a brisk tug.