129 Balcombe Close
In town the earliest risers were just beginning to look sleepily from their windows as we drove
through the streets of the Surrey side. Passing down the Waterloo Bridge Road we crossed over the
river, and dashing up Wellington Street wheeled sharply to the right and found ourselves in Bow
Street. Sherlock Holmes was well known to the force, and the two constables at the door saluted him.
One of them held the horse's head while the other led us in.

"Who is on duty?" asked Holmes.

"Inspector Bradstreet, sir."