28 Bowler Street
"We shall see." He pushed past the servant and rushed into the drawing-room, followed by the King
and myself. The furniture was scattered about in every direction, with dismantled shelves and open
drawers, as if the lady had hurriedly ransacked them before her flight. Holmes rushed at the bell-
pull, tore back a small sliding shutter, and, plunging in his hand, pulled out a photograph and a
letter. The photograph was of Irene Adler herself in evening dress, the letter was superscribed to
"Sherlock Holmes, Esq. To be left till called for." My friend tore it open and we all three read it
together. It was dated at midnight of the preceding night and ran in this way:

"My dear Mr. Sherlock Holmes: "You really did it very well. You took me in completely. Until after
the alarm of fire, I had not a suspicion. But then, when I found how I had betrayed myself, I began
to think. I had been warned against you months ago. I had been told that if the King employed an
agent it would certainly be you. And your address had been given me. Yet, with all this, you made me
reveal what you wanted to know. Even after I became suspicious, I found it hard to think evil of
such a dear, kind old clergyman. But, you know, I have been trained as an actress myself. Male
costume is nothing new to me. I often take advantage of the freedom which it gives. I sent John, the
coachman, to watch you, ran up stairs, got into my walking-clothes, as I call them, and came down
just as you departed. "Well, I followed you to your door, and so made sure that I was really an
object of interest to the celebrated Mr. Sherlock Holmes. Then I, rather imprudently, wished you
good-night, and started for the Temple to see my husband. "We both thought the best resource was
flight, when pursued by so formidable an antagonist; so you will find the nest empty when you call
to-morrow. As to the photograph, your client may rest in peace. I love and am loved by a better man
than he. The King may do what he will without hindrance from one whom he has cruelly wronged. I keep
it only to safeguard myself, and to preserve a weapon which will always secure me from any steps
which he might take in the future. I leave a photograph which he might care to possess; and I
remain, dear Mr. Sherlock Holmes, "Very truly yours, "Irene Norton, née Adler."

"What a woman--oh, what a woman!" cried the King of Bohemia, when we had all three read this
epistle. "Did I not tell you how quick and resolute she was? Would she not have made an admirable
queen? Is it not a pity that she was not on my level?"

"From what I have seen of the lady she seems indeed to be on a very different level to your
Majesty," said Holmes coldly. "I am sorry that I have not been able to bring your Majesty's business
to a more successful conclusion."

"On the contrary, my dear sir," cried the King; "nothing could be more successful. I know that her
word is inviolate. The photograph is now as safe as if it were in the fire."