42 Corbet Close
"You are engaged," said I; "perhaps I interrupt you."

"Not at all. I am glad to have a friend with whom I can discuss my results. The matter is a
perfectly trivial one"--he jerked his thumb in the direction of the old hat--"but there are points
in connection with it which are not entirely devoid of interest and even of instruction."