35 Moor End
It was early in April in the year '83 that I woke one morning to find Sherlock Holmes standing,
fully dressed, by the side of my bed. He was a late riser, as a rule, and as the clock on the
mantelpiece showed me that it was only a quarter-past seven, I blinked up at him in some surprise,
and perhaps just a little resentment, for I was myself regular in my habits.

"Very sorry to knock you up, Watson," said he, "but it's the common lot this morning. Mrs. Hudson
has been knocked up, she retorted upon me, and I on you."