137 Wantage Close
"Mr. Hosmer Angel came to the house again and proposed that we should marry before father came back.
He was in dreadful earnest and made me swear, with my hands on the Testament, that whatever happened
I would always be true to him. Mother said he was quite right to make me swear, and that it was a
sign of his passion. Mother was all in his favour from the first and was even fonder of him than I
was. Then, when they talked of marrying within the week, I began to ask about father; but they both
said never to mind about father, but just to tell him afterwards, and mother said she would make it
all right with him. I didn't quite like that, Mr. Holmes. It seemed funny that I should ask his
leave, as he was only a few years older than me; but I didn't want to do anything on the sly, so I
wrote to father at Bordeaux, where the company has its French offices, but the letter came back to
me on the very morning of the wedding."

"It missed him, then?"

"Yes, sir; for he had started to England just before it arrived."

"Ha! that was unfortunate. Your wedding was arranged, then, for the Friday. Was it to be in church?"

"Yes, sir, but very quietly. It was to be at St. Saviour's, near King's Cross, and we were to have
breakfast afterwards at the St. Pancras Hotel. Hosmer came for us in a hansom, but as there were two
of us he put us both into it and stepped himself into a four-wheeler, which happened to be the only
other cab in the street. We got to the church first, and when the four-wheeler drove up we waited
for him to step out, but he never did, and when the cabman got down from the box and looked there
was no one there! The cabman said that he could not imagine what had become of him, for he had seen
him get in with his own eyes. That was last Friday, Mr. Holmes, and I have never seen or heard
anything since then to throw any light upon what became of him."