37 Balcombe Close
He picked it up and gazed at it in the peculiar introspective fashion which was characteristic of
him. "It is perhaps less suggestive than it might have been," he remarked, "and yet there are a few
inferences which are very distinct, and a few others which represent at least a strong balance of
probability. That the man was highly intellectual is of course obvious upon the face of it, and also
that he was fairly well-to-do within the last three years, although he has now fallen upon evil
days. He had foresight, but has less now than formerly, pointing to a moral retrogression, which,
when taken with the decline of his fortunes, seems to indicate some evil influence, probably drink,
at work upon him. This may account also for the obvious fact that his wife has ceased to love him."

"My dear Holmes!"

"He has, however, retained some degree of self-respect," he continued, disregarding my remonstrance.
"He is a man who leads a sedentary life, goes out little, is out of training entirely, is middle-
aged, has grizzled hair which he has had cut within the last few days, and which he anoints with
lime-cream. These are the more patent facts which are to be deduced from his hat. Also, by the way,
that it is extremely improbable that he has gas laid on in his house."

"You are certainly joking, Holmes."

"Not in the least. Is it possible that even now, when I give you these results, you are unable to
see how they are attained?"