61 Bowler Street
"When did it break out?"

"I hear that it was during the night, sir, but it has got worse, and the whole place is in a blaze."

"Whose house is it?"

"Dr. Becher's."

"Tell me," broke in the engineer, "is Dr. Becher a German, very thin, with a long, sharp nose?"