147 Wantage Close
"'And be laughed at for my pains. Nothing of the sort.'

"'Then let me do so?'

"'No, I forbid you. I won't have a fuss made about such nonsense.'

"It was in vain to argue with him, for he was a very obstinate man. I went about, however, with a
heart which was full of forebodings.

"On the third day after the coming of the letter my father went from home to visit an old friend of
his, Major Freebody, who is in command of one of the forts upon Portsdown Hill. I was glad that he
should go, for it seemed to me that he was farther from danger when he was away from home. In that,
however, I was in error. Upon the second day of his absence I received a telegram from the major,
imploring me to come at once. My father had fallen over one of the deep chalk-pits which abound in
the neighbourhood, and was lying senseless, with a shattered skull. I hurried to him, but he passed
away without having ever recovered his consciousness. He had, as it appears, been returning from
Fareham in the twilight, and as the country was unknown to him, and the chalk-pit unfenced, the jury
had no hesitation in bringing in a verdict of 'death from accidental causes.' Carefully as I
examined every fact connected with his death, I was unable to find anything which could suggest the
idea of murder. There were no signs of violence, no footmarks, no robbery, no record of strangers
having been seen upon the roads. And yet I need not tell you that my mind was far from at ease, and
that I was well-nigh certain that some foul plot had been woven round him.

"In this sinister way I came into my inheritance. You will ask me why I did not dispose of it? I
answer, because I was well convinced that our troubles were in some way dependent upon an incident
in my uncle's life, and that the danger would be as pressing in one house as in another.