175 Down Close
"Yes, I have got it now," he answered with his thick red finger planted halfway down the column.
"Here it is. This is what began it all. You just read it for yourself, sir."

I took the paper from him and read as follows:

"To the Red-headed League: On account of the bequest of the late Ezekiah Hopkins, of Lebanon,
Pennsylvania, U. S. A., there is now another vacancy open which entitles a member of the League to a
salary of £4 a week for purely nominal services. All red-headed men who are sound in body and mind
and above the age of twenty-one years, are eligible. Apply in person on Monday, at eleven o'clock,
to Duncan Ross, at the offices of the League, 7 Pope's Court, Fleet Street."