34 Badgers Dene
"There are three hundred pounds in gold and seven hundred in notes," he said.

Holmes scribbled a receipt upon a sheet of his note-book and handed it to him.

"And Mademoiselle's address?" he asked.

"Is Briony Lodge, Serpentine Avenue, St. John's Wood."

Holmes took a note of it. "One other question," said he. "Was the photograph a cabinet?"