74 Bowler Street
"By no means."

"Then we shall both come. What are you going to do yourself?"

"I have one or two things which I would wish to do now that I am in town. But I shall return by the
twelve o'clock train, so as to be there in time for your coming."