10 Ventnor Terrace
"Well, very soon I found that I was saving considerable sums of money. I do not mean that any beggar
in the streets of London could earn £700 a year--which is less than my average takings--but I had
exceptional advantages in my power of making up, and also in a facility of repartee, which improved
by practice and made me quite a recognised character in the City. All day a stream of pennies,
varied by silver, poured in upon me, and it was a very bad day in which I failed to take £2.

"As I grew richer I grew more ambitious, took a house in the country, and eventually married,
without anyone having a suspicion as to my real occupation. My dear wife knew that I had business in
the City. She little knew what.

"Last Monday I had finished for the day and was dressing in my room above the opium den when I
looked out of my window and saw, to my horror and astonishment, that my wife was standing in the
street, with her eyes fixed full upon me. I gave a cry of surprise, threw up my arms to cover my
face, and, rushing to my confidant, the Lascar, entreated him to prevent anyone from coming up to
me. I heard her voice downstairs, but I knew that she could not ascend. Swiftly I threw off my
clothes, pulled on those of a beggar, and put on my pigments and wig. Even a wife's eyes could not
pierce so complete a disguise. But then it occurred to me that there might be a search in the room,
and that the clothes might betray me. I threw open the window, reopening by my violence a small cut
which I had inflicted upon myself in the bedroom that morning. Then I seized my coat, which was
weighted by the coppers which I had just transferred to it from the leather bag in which I carried
my takings. I hurled it out of the window, and it disappeared into the Thames. The other clothes
would have followed, but at that moment there was a rush of constables up the stair, and a few
minutes after I found, rather, I confess, to my relief, that instead of being identified as Mr.
Neville St. Clair, I was arrested as his murderer.