30 Cliff Place
"Precisely. And if it were guilty, why did he not invent a lie? His silence appears to me to cut
both ways. There are several singular points about the case. What did the police think of the noise
which awoke you from your sleep?"

"They considered that it might be caused by Arthur's closing his bedroom door."

"A likely story! As if a man bent on felony would slam his door so as to wake a household. What did
they say, then, of the disappearance of these gems?"

"They are still sounding the planking and probing the furniture in the hope of finding them."

"Have they thought of looking outside the house?"