147 Glendale
"If you remember, Monday was an exceedingly hot day, and Mrs. St. Clair walked slowly, glancing
about in the hope of seeing a cab, as she did not like the neighbourhood in which she found herself.
While she was walking in this way down Swandam Lane, she suddenly heard an ejaculation or cry, and
was struck cold to see her husband looking down at her and, as it seemed to her, beckoning to her
from a second-floor window. The window was open, and she distinctly saw his face, which she
describes as being terribly agitated. He waved his hands frantically to her, and then vanished from
the window so suddenly that it seemed to her that he had been plucked back by some irresistible
force from behind. One singular point which struck her quick feminine eye was that although he wore
some dark coat, such as he had started to town in, he had on neither collar nor necktie.

"Convinced that something was amiss with him, she rushed down the steps--for the house was none
other than the opium den in which you found me to-night--and running through the front room she
attempted to ascend the stairs which led to the first floor. At the foot of the stairs, however, she
met this Lascar scoundrel of whom I have spoken, who thrust her back and, aided by a Dane, who acts
as assistant there, pushed her out into the street. Filled with the most maddening doubts and fears,
she rushed down the lane and, by rare good-fortune, met in Fresno Street a number of constables with
an inspector, all on their way to their beat. The inspector and two men accompanied her back, and in
spite of the continued resistance of the proprietor, they made their way to the room in which Mr.
St. Clair had last been seen. There was no sign of him there. In fact, in the whole of that floor
there was no one to be found save a crippled wretch of hideous aspect, who, it seems, made his home
there. Both he and the Lascar stoutly swore that no one else had been in the front room during the
afternoon. So determined was their denial that the inspector was staggered, and had almost come to
believe that Mrs. St. Clair had been deluded when, with a cry, she sprang at a small deal box which
lay upon the table and tore the lid from it. Out there fell a cascade of children's bricks. It was
the toy which he had promised to bring home.