76 Bowler Street
"That will do, Mr. Wilson. I shall be happy to give you an opinion upon the subject in the course of
a day or two. To-day is Saturday, and I hope that by Monday we may come to a conclusion."

"Well, Watson," said Holmes when our visitor had left us, "what do you make of it all?"

"I make nothing of it," I answered frankly. "It is a most mysterious business."