54 Beaconside
It was a quarter past six when we left Baker Street, and it still wanted ten minutes to the hour
when we found ourselves in Serpentine Avenue. It was already dusk, and the lamps were just being
lighted as we paced up and down in front of Briony Lodge, waiting for the coming of its occupant.
The house was just such as I had pictured it from Sherlock Holmes' succinct description, but the
locality appeared to be less private than I expected. On the contrary, for a small street in a quiet
neighbourhood, it was remarkably animated. There was a group of shabbily dressed men smoking and
laughing in a corner, a scissors-grinder with his wheel, two guardsmen who were flirting with a
nurse-girl, and several well-dressed young men who were lounging up and down with cigars in their
mouths.

"You see," remarked Holmes, as we paced to and fro in front of the house, "this marriage rather
simplifies matters. The photograph becomes a double-edged weapon now. The chances are that she would
be as averse to its being seen by Mr. Godfrey Norton, as our client is to its coming to the eyes of
his princess. Now the question is--Where are we to find the photograph?"

"Where, indeed?"

"It is most unlikely that she carries it about with her. It is cabinet size. Too large for easy
concealment about a woman's dress. She knows that the King is capable of having her waylaid and
searched. Two attempts of the sort have already been made. We may take it, then, that she does not
carry it about with her."