162 Wantage Close
"You may place considerable confidence in Mr. Holmes, sir," said the police agent loftily. "He has
his own little methods, which are, if he won't mind my saying so, just a little too theoretical and
fantastic, but he has the makings of a detective in him. It is not too much to say that once or
twice, as in that business of the Sholto murder and the Agra treasure, he has been more nearly
correct than the official force."

"Oh, if you say so, Mr. Jones, it is all right," said the stranger with deference. "Still, I confess
that I miss my rubber. It is the first Saturday night for seven-and-twenty years that I have not had
my rubber."

"I think you will find," said Sherlock Holmes, "that you will play for a higher stake to-night than
you have ever done yet, and that the play will be more exciting. For you, Mr. Merryweather, the
stake will be some £30,000; and for you, Jones, it will be the man upon whom you wish to lay your
hands."