141 Glendale
"It is to him that this trophy belongs."

"It is his hat."

"No, no, he found it. Its owner is unknown. I beg that you will look upon it not as a battered
billycock but as an intellectual problem. And, first, as to how it came here. It arrived upon
Christmas morning, in company with a good fat goose, which is, I have no doubt, roasting at this
moment in front of Peterson's fire. The facts are these: about four o'clock on Christmas morning,
Peterson, who, as you know, is a very honest fellow, was returning from some small jollification and
was making his way homeward down Tottenham Court Road. In front of him he saw, in the gaslight, a
tallish man, walking with a slight stagger, and carrying a white goose slung over his shoulder. As
he reached the corner of Goodge Street, a row broke out between this stranger and a little knot of
roughs. One of the latter knocked off the man's hat, on which he raised his stick to defend himself
and, swinging it over his head, smashed the shop window behind him. Peterson had rushed forward to
protect the stranger from his assailants; but the man, shocked at having broken the window, and
seeing an official-looking person in uniform rushing towards him, dropped his goose, took to his
heels, and vanished amid the labyrinth of small streets which lie at the back of Tottenham Court
Road. The roughs had also fled at the appearance of Peterson, so that he was left in possession of
the field of battle, and also of the spoils of victory in the shape of this battered hat and a most
unimpeachable Christmas goose."

"Which surely he restored to their owner?"