92 Wantage Close
"Boscombe Valley is a country district not very far from Ross, in Herefordshire. The largest landed
proprietor in that part is a Mr. John Turner, who made his money in Australia and returned some
years ago to the old country. One of the farms which he held, that of Hatherley, was let to Mr.
Charles McCarthy, who was also an ex-Australian. The men had known each other in the colonies, so
that it was not unnatural that when they came to settle down they should do so as near each other as
possible. Turner was apparently the richer man, so McCarthy became his tenant but still remained, it
seems, upon terms of perfect equality, as they were frequently together. McCarthy had one son, a lad
of eighteen, and Turner had an only daughter of the same age, but neither of them had wives living.
They appear to have avoided the society of the neighbouring English families and to have led retired
lives, though both the McCarthys were fond of sport and were frequently seen at the race-meetings of
the neighbourhood. McCarthy kept two servants--a man and a girl. Turner had a considerable
household, some half-dozen at the least. That is as much as I have been able to gather about the
families. Now for the facts.

"On June 3rd, that is, on Monday last, McCarthy left his house at Hatherley about three in the
afternoon and walked down to the Boscombe Pool, which is a small lake formed by the spreading out of
the stream which runs down the Boscombe Valley. He had been out with his serving-man in the morning
at Ross, and he had told the man that he must hurry, as he had an appointment of importance to keep
at three. From that appointment he never came back alive.