150 Down Close
"But the twelve-mile drive?" gasped Hatherley.

"Six out and six back. Nothing simpler. You say yourself that the horse was fresh and glossy when
you got in. How could it be that if it had gone twelve miles over heavy roads?"

"Indeed, it is a likely ruse enough," observed Bradstreet thoughtfully. "Of course there can be no
doubt as to the nature of this gang."