122 Corbet Close
"I was afraid that you were engaged."

"So I am. Very much so."

"Then I can wait in the next room."

"Not at all. This gentleman, Mr. Wilson, has been my partner and helper in many of my most
successful cases, and I have no doubt that he will be of the utmost use to me in yours also."