20 Glan Road
"My dear fellow, I know you well. I know the military neatness which characterises you. You shave
every morning, and in this season you shave by the sunlight; but since your shaving is less and less
complete as we get farther back on the left side, until it becomes positively slovenly as we get
round the angle of the jaw, it is surely very clear that that side is less illuminated than the
other. I could not imagine a man of your habits looking at himself in an equal light and being
satisfied with such a result. I only quote this as a trivial example of observation and inference.
Therein lies my métier, and it is just possible that it may be of some service in the investigation
which lies before us. There are one or two minor points which were brought out in the inquest, and
which are worth considering."

"What are they?"