124 Valentia Road
"Give me a pencil and that slip of paper. Now, then: 'Found at the corner of Goodge Street, a goose
and a black felt hat. Mr. Henry Baker can have the same by applying at 6.30 this evening at 221b,
Baker Street.' That is clear and concise."

"Very. But will he see it?"