109 Larch Walk
"'I should certainly speak to the police,' I said.

"'And be laughed at for my pains. Nothing of the sort.'

"'Then let me do so?'

"'No, I forbid you. I won't have a fuss made about such nonsense.'

"It was in vain to argue with him, for he was a very obstinate man. I went about, however, with a
heart which was full of forebodings.