13 Bowler Street
"I didn't know what to do, so I came straight to you." That was always the way. Folk who were in
grief came to my wife like birds to a light-house.

"It was very sweet of you to come. Now, you must have some wine and water, and sit here comfortably
and tell us all about it. Or should you rather that I sent James off to bed?"

"Oh, no, no! I want the doctor's advice and help, too. It's about Isa. He has not been home for two
days. I am so frightened about him!"

It was not the first time that she had spoken to us of her husband's trouble, to me as a doctor, to
my wife as an old friend and school companion. We soothed and comforted her by such words as we
could find. Did she know where her husband was? Was it possible that we could bring him back to her?

It seems that it was. She had the surest information that of late he had, when the fit was on him,
made use of an opium den in the farthest east of the City. Hitherto his orgies had always been
confined to one day, and he had come back, twitching and shattered, in the evening. But now the
spell had been upon him eight-and-forty hours, and he lay there, doubtless among the dregs of the
docks, breathing in the poison or sleeping off the effects. There he was to be found, she was sure
of it, at the Bar of Gold, in Upper Swandam Lane. But what was she to do? How could she, a young and
timid woman, make her way into such a place and pluck her husband out from among the ruffians who
surrounded him?

There was the case, and of course there was but one way out of it. Might I not escort her to this
place? And then, as a second thought, why should she come at all? I was Isa Whitney's medical
adviser, and as such I had influence over him. I could manage it better if I were alone. I promised
her on my word that I would send him home in a cab within two hours if he were indeed at the address
which she had given me. And so in ten minutes I had left my armchair and cheery sitting-room behind
me, and was speeding eastward in a hansom on a strange errand, as it seemed to me at the time,
though the future only could show how strange it was to be.