2 Wellswood Gardens
"What, then?"

"It is fear, Mr. Holmes. It is terror." She raised her veil as she spoke, and we could see that she
was indeed in a pitiable state of agitation, her face all drawn and grey, with restless frightened
eyes, like those of some hunted animal. Her features and figure were those of a woman of thirty, but
her hair was shot with premature grey, and her expression was weary and haggard. Sherlock Holmes ran
her over with one of his quick, all-comprehensive glances.