22 Balcombe Close
"Why," said I, glancing up at my companion, "that was surely the bell. Who could come to-night? Some
friend of yours, perhaps?"

"Except yourself I have none," he answered. "I do not encourage visitors."

"A client, then?"

"If so, it is a serious case. Nothing less would bring a man out on such a day and at such an hour.
But I take it that it is more likely to be some crony of the landlady's."

Sherlock Holmes was wrong in his conjecture, however, for there came a step in the passage and a
tapping at the door. He stretched out his long arm to turn the lamp away from himself and towards
the vacant chair upon which a newcomer must sit.

"Come in!" said he.