51 Glendale
"A house on fire?" asked Bradstreet as the train steamed off again on its way.

"Yes, sir!" said the station-master.

"When did it break out?"

"I hear that it was during the night, sir, but it has got worse, and the whole place is in a blaze."

"Whose house is it?"